Adhoc build 1.0.0.20131207_164840

+++ PseudoTerminal.m
@@ -266,7 +266,7 @@ NSString *sessionsKey = @"sessions";
             DLog(@"Convert default screen to screen number: No screen matches the window's screen so using main screen");
             screenNumber = 0;
         } else {
-            DLog(@"Convert default screen to screen number: System chose screen %d", n);
+            DLog(@"Convert default screen to screen number: System chose screen %lu", (unsigned long)n);
             screenNumber = n;
         }
     }
